KeyGen: indifferentiability for HKDF
We can prove indifferentiability for HKDF just if we can be sure
that IKM ends with an octet that's different from the final octet
of the input to any invocation of HMAC inside of HKDF-Expand.
Fortunately, HKDF-Expand's counter starts at 1 and is appended
to the input to HMAC---so we can get indifferentiability just
by appending a 0 octet to IKM.

This is more of a safety / paranoia thing than anything else.

Also, the prior commit added key_info to optionally orthogonalize
different invocations of KeyGen called on the same key. In this
commit, I also appended a two-byte encoding of L to the info field.
This means that using the same keying material to extract different
lengths of key will give different results, which seems nice.